pyproject.toml
[project]
name = "Amaya"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "python-telegram-bot",
    "loguru",
    "python-dotenv",
    "pyee",
    "openai",
    "aiosqlite",
    "watchdog",
    "python-ulid",
    "pytz",
]
